Make ClusterConsumer interface private (clusterConsumer)
